requirements.txt
amqp==2.5.2
appdirs==1.4.4
attrs==19.3.0
backcall==0.1.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
click==7.1.2
decorator==4.4.2
Django==2.1.10
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
fpdf==1.7.2
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kombu==4.6.8
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
pytesseract==0.3.4
pytz==2020.1
redis==3.5.2
regex==2020.5.14
six==1.14.0
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
vine==1.3.0
wcwidth==0.1.9
zipp==3.1.0
